[SPARK-31748][PYTHON][3.0] Document resource module in PySpark doc and rename/move classes
### What changes were proposed in this pull request?

This PR partially backports https://github.com/apache/spark/pull/28569.

1.. Rename

```
pyspark
└── resourceinformation.py
    └── class ResourceInformation
```

to

```
pyspark
└── resource.py
    └── class ResourceInformation
```

So, the `ResourceInformation` is consistently imported via `pyspark.resource.ResourceInformation`.

2.. Document the new `pyspark.resource` module

3.. Minor docstring fix e.g.:

```diff
-     param name the name of the resource
-     param addresses an array of strings describing the addresses of the resource
+     :param name: the name of the resource
+     :param addresses: an array of strings describing the addresses of the resource
+
+     .. versionadded:: 3.0.0
```

### Why are the changes needed?

To document APIs, and move Python modules to fewer and simpler modules.

### Does this PR introduce _any_ user-facing change?

No, the changes are in unreleased branches.

### How was this patch tested?

Manually tested via:

```bash
cd python
./run-tests --python-executables=python3 --modules=pyspark-core
```

Closes #28586 from HyukjinKwon/SPARK-31748.

Authored-by: HyukjinKwon <gurwls223@apache.org>
